+++ mipac/actions/mute.py
@@ -1,8 +1,12 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, AsyncIterable
 
 from mipac.abstract.action import AbstractAction
+from mipac.errors.base import ParameterError
 from mipac.http import HTTPClient, Route
+from mipac.models.mute import MuteUser
+from mipac.types.mute import IMuteUser
+from mipac.util import remove_dict_empty
 
 if TYPE_CHECKING:
     from mipac.manager.client import ClientActions
@@ -63,3 +67,44 @@ class MuteActions(AbstractAction):
             Route('POST', '/api/mute/delete'), auth=True, json=data
         )
         return res
+
+    async def gets(
+        self,
+        limit: int = 100,
+        since_id: str | None = None,
+        until_id: str | None = None,
+        all: bool = True,
+    ) -> AsyncIterable[MuteUser]:
+        if limit > 100:
+            raise ParameterError('limit は100以下である必要があります')
+
+        async def request(body) -> list[MuteUser]:
+            res: list[IMuteUser] = await self._session.request(
+                Route('POST', '/api/mute/list'),
+                lower=True,
+                auth=True,
+                json=body,
+            )
+            return [MuteUser(user, client=self._client) for user in res]
+
+        data = remove_dict_empty(
+            {'limit': limit, 'sinceId': since_id, 'untilId': until_id}
+        )
+
+        if all:
+            data['limit'] = 100
+        first_req = await request(data)
+
+        for user in first_req:
+            yield user
+
+        if all and len(first_req) == 100:
+            data['untilId'] = first_req[-1].id
+            while True:
+                res = await request(data)
+                if len(res) <= 100:
+                    for user in res:
+                        yield user
+                if len(res) == 0:
+                    break
+                data['untilId'] = res[-1].id